move dry run to src/server package
